requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
APScheduler
pyspark
kafka-python
cassandra-driver
